# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SecuritySuite (secapi) — APIs propias",
    "version": "0.1.0",
    "description": "Catálogo generado por `pnpm docs:api` a partir de `src/app/api/**/route.ts`.\n\nLo generado describe **lo que el código hace hoy**, incluida la autenticación real de\ncada handler. Las descripciones a mano viven en `docs/api/anotaciones.yaml` y ganan\nsobre lo inferido; `GET /api/docs/spec` devuelve el documento ya mergeado.\n\nNo editar este archivo a mano: se regenera."
  },
  "servers": [
    {
      "url": "http://localhost:4005",
      "description": "desarrollo local (pnpm dev)"
    },
    {
      "url": "https://secapi.riogas.com.uy",
      "description": "secapi (PM2, puerto 3001)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerJwt": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT emitido por `POST /api/db/login`. Payload: `{ iss, username, userId, sistema }` — no lleva el flag root."
      },
      "cookieToken": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token",
        "description": "El mismo JWT, tomado de la cookie `token` cuando no hay header."
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    }
  },
  "paths": {
...
    "/api/db/roles/{id}/atributos": {
      "get": {
        "tags": [
          "roles"
        ],
        "summary": "GET /api/db/roles/[id]/atributos — atributos del rol",
        "description": "GET /api/db/roles/[id]/atributos — atributos del rol",
        "operationId": "get_api_db_roles_id_atributos",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Segmento dinámico `[id]` de la ruta."
          }
        ],
        "responses": {
          "default": {
            "description": "Sin anotar todavía (ver docs/api/README.md)."
          }
        },
        "x-auth": {
          "modo": "jwt",
          "verificaFirma": true,
          "detectadoEn": "handler",
          "detalle": "JWT en `Authorization: Bearer` o cookie `token`, con firma y vencimiento verificados por el guard central de `/api/db` (`requireApiAuth`, src/lib/auth/apiGuard.ts). El nivel exigido por esta ruta está declarado en `POLITICAS`, en ese mismo archivo."
        },
        "x-archivo": "src/app/api/db/roles/[id]/atributos/route.ts",
        "x-generado": true,
        "security": [
          {
            "bearerJwt": []
          },
          {
            "cookieToken": []
          }
        ]
      },
      "post": {
        "tags": [
          "roles"
        ],
        "summary": "POST /api/db/roles/[id]/atributos — agregar o actualizar un único atributo",
        "description": "POST /api/db/roles/[id]/atributos — agregar o actualizar un único atributo",
        "operationId": "post_api_db_roles_id_atributos",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Segmento dinámico `[id]` de la ruta."
          }
        ],
        "responses": {
          "default": {
            "description": "Sin anotar todavía (ver docs/api/README.md)."
          }
        },
        "x-auth": {
          "modo": "jwt",
          "verificaFirma": true,
          "detectadoEn": "handler",
          "detalle": "JWT en `Authorization: Bearer` o cookie `token`, con firma y vencimiento verificados por el guard central de `/api/db` (`requireApiAuth`, src/lib/auth/apiGuard.ts). El nivel exigido por esta ruta está declarado en `POLITICAS`, en ese mismo archivo."
        },
        "x-archivo": "src/app/api/db/roles/[id]/atributos/route.ts",
        "x-generado": true,
        "security": [
          {
            "bearerJwt": []
          },
          {
            "cookieToken": []
          }
        ]
      },
      "put": {
        "tags": [
          "roles"
        ],
        "summary": "PUT /api/db/roles/[id]/atributos — reemplaza todos los atributos del rol",
        "description": "PUT /api/db/roles/[id]/atributos — reemplaza todos los atributos del rol\nBody: { atributos: [{ atributo, valor }] }",
        "operationId": "put_api_db_roles_id_atributos",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Segmento dinámico `[id]` de la ruta."
          }
        ],
        "responses": {
          "default": {
            "description": "Sin anotar todavía (ver docs/api/README.md)."
          }
        },
        "x-auth": {
          "modo": "jwt",
          "verificaFirma": true,
          "detectadoEn": "handler",
          "detalle": "JWT en `Authorization: Bearer` o cookie `token`, con firma y vencimiento verificados por el guard central de `/api/db` (`requireApiAuth`, src/lib/auth/apiGuard.ts). El nivel exigido por esta ruta está declarado en `POLITICAS`, en ese mismo archivo."
        },
        "x-archivo": "src/app/api/db/roles/[id]/atributos/route.ts",
        "x-generado": true,
        "security": [
          {
            "bearerJwt": []
          },
          {
            "cookieToken": []
          }
        ]
      }
    },
...
    "/api/db/usuarios/{id}/accesos": {
      "get": {
        "tags": [
          "usuarios"
        ],
        "summary": "Devuelve todas las funcionalidades del sistema con:",
        "description": "GET /api/db/usuarios/[id]/accesos\nDevuelve todas las funcionalidades del sistema con:\n  - accesoDirecto: override en tabla `accesos` (grant/deny) o null\n  - viaRoles: nombres de roles del usuario que la otorgan",
        "operationId": "get_api_db_usuarios_id_accesos",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Segmento dinámico `[id]` de la ruta."
          }
        ],
        "responses": {
          "default": {
            "description": "Sin anotar todavía (ver docs/api/README.md)."
          }
        },
        "x-auth": {
          "modo": "jwt",
          "verificaFirma": true,
          "detectadoEn": "handler",
          "detalle": "JWT en `Authorization: Bearer` o cookie `token`, con firma y vencimiento verificados por el guard central de `/api/db` (`requireApiAuth`, src/lib/auth/apiGuard.ts). El nivel exigido por esta ruta está declarado en `POLITICAS`, en ese mismo archivo."
        },
        "x-archivo": "src/app/api/db/usuarios/[id]/accesos/route.ts",
        "x-generado": true,
        "security": [
          {
            "bearerJwt": []
          },
          {
            "cookieToken": []
          }
        ]
      },
      "put": {
        "tags": [
          "usuarios"
        ],
        "summary": "Bulk upsert/delete de accesos directos.",
        "description": "PUT /api/db/usuarios/[id]/accesos\nBulk upsert/delete de accesos directos.\nBody: { cambios: CambioAcceso[] }",
        "operationId": "put_api_db_usuarios_id_accesos",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Segmento dinámico `[id]` de la ruta."
          }
        ],
        "responses": {
          "default": {
            "description": "Sin anotar todavía (ver docs/api/README.md)."
          }
        },
        "x-auth": {
          "modo": "jwt",
          "verificaFirma": true,
          "detectadoEn": "handler",
          "detalle": "JWT en `Authorization: Bearer` o cookie `token`, con firma y vencimiento verificados por el guard central de `/api/db` (`requireApiAuth`, src/lib/auth/apiGuard.ts). El nivel exigido por esta ruta está declarado en `POLITICAS`, en ese mismo archivo."
        },
        "x-archivo": "src/app/api/db/usuarios/[id]/accesos/route.ts",
        "x-generado": true,
        "security": [
          {
            "bearerJwt": []
          },
          {
            "cookieToken": []
          }
        ]
      }
    },
...
  },
...
}